fast: add the FAST 12 laureates and make it dynamic
Add the FAST information directly on the startup Frontmatter (i.e the
YAML data) and make the FAST 12 page use it via a new dynamic partial
that's included if the "new-fast" page attribute is true.

This means that in the future the FAST team can do it all themselves,
for example with the next edition:

1. add a new 13_fast.md page that has a `new-fast: true` property ;
2. go through the startups that won and add a `fast` YML prop with the
`promotion`, `amount`, `ambition` and `impact` properties in there.

And then it just works™